poc: per-vhost connection_created metric
I've provided a new metric family, the `connection_created` one under
the detailed group of metrics. this metric is intended to allow querying
the current number of connections per vhost.

Garbage collection mechanism for this metric needs to be provided, but
provided source seems enought as to evalue if we should keep building
a solution that could be integrated in RabbitMQ.