Fix QQ periodic policy repaire when delivery-limit is disabled
In queue args and policies delivery-limit can be disabled by setting a
negative value. `rabbit_fifo` converts this to undefined so when
`maybe_apply_policies` compares current config with new config, the
former contains `undefined` while the later contains a negative value
for the same unchanged "disabled" delivery limit.

This commit converts a negative value to `undefined` when comparig
configs.